Fix: 관심사 추가 API 수정

+++ src/main/java/Lingtning/new_match42/dto/UserResponse.java
@@ -17,8 +17,8 @@ public class UserResponse {
     private Role role;
 
     private List<String> interests;
-    private Integer interestCount;
+    private Long interestCount;
 
     private List<String> blockUsers;
-    private Integer blockCount;
+    private Long blockCount;
 }
